Move IndexGatewayClientCfg to gateway_client file.
